requirements.txt
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.13.0
astroid==4.0.4
certifi==2026.2.25
charset-normalizer==3.4.6
coverage==7.13.5
dill==0.4.1
distro==1.9.0
docstring_parser==0.17.0
et_xmlfile==2.0.0
execnet==2.1.2
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
isort==8.0.1
jiter==0.13.0
mccabe==0.7.0
openai==2.30.0
openpyxl==3.1.5
packaging==26.0
platformdirs==4.9.4
playwright==1.58.0
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
Pygments==2.19.2
pylint==4.0.5
pytest==9.0.2
pytest-base-url==2.1.0
pytest-cov==7.1.0
pytest-playwright==0.7.2
pytest-xdist==3.8.0
python-dotenv==1.2.2
python-slugify==8.0.4
PyYAML==6.0.3
redis==7.4.0
requests==2.33.0
sniffio==1.3.1
structlog==25.5.0
tenacity==9.1.4
text-unidecode==1.3
tomlkit==0.14.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
